WIP MIDIplay v2
