+++ src/midiplay/event_hooks.hpp
@@ -1,0 +1,48 @@
+/*
+ * libADLMIDI is a free Software MIDI synthesizer library with OPL3 emulation
+ *
+ * ADLMIDI Library API:   Copyright (c) 2015-2019 Vitaly Novichkov <admin@wohlnet.ru>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef OPLMIDI_MIDIPLAY_EVENT_HOOKS_HPP
+#define OPLMIDI_MIDIPLAY_EVENT_HOOKS_HPP
+
+#include <cstddef>
+
+/**
+ * @brief Hooks of the internal events
+ */
+struct MIDIEventHooks
+{
+    MIDIEventHooks() :
+        onNote(NULL),
+        onNote_userData(NULL),
+        onDebugMessage(NULL),
+        onDebugMessage_userData(NULL)
+    {}
+
+    //! Note on/off hooks
+    typedef void (*NoteHook)(void *userdata, int adlchn, int note, int ins, int pressure, double bend);
+    NoteHook     onNote;
+    void         *onNote_userData;
+
+    //! Library internal debug messages
+    typedef void (*DebugMessageHook)(void *userdata, const char *fmt, ...);
+    DebugMessageHook onDebugMessage;
+    void *onDebugMessage_userData;
+};
+
+#endif // OPLMIDI_MIDIPLAY_EVENT_HOOKS_HPP
